module1/demo/prepared/config-service/pyproject.toml
[project]
name = "config-service"
version = "0.1.0"
description = "REST Web API for managing application configurations"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.116.1",
    "pydantic==2.11.7",
    "pydantic-settings>=2.0.0,<3.0.0",
    "pydantic-extra-types>=2.0.0,<3.0.0",
    "python-ulid>=2.0.0,<3.0.0",
    "psycopg2-binary==2.9.10",
    "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
test = [
    "pytest==8.4.1",
    "httpx==0.28.1",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["config_service"]
python_files = ["*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
